add permission

+++ app/Http/Controllers/ExchangeController.php
@@ -13,20 +13,36 @@ class ExchangeController extends Controller
     }
     public function index()
     {
+        if(!Right::check('Exchange', 'l'))
+        {
+            return view('admins.permissions.no');
+        }
         $data['excs'] = DB::table('exchanges')->paginate(18);
         return view('admins.exchanges.index', $data);
     }
     public function create()
     {
+        if(!Right::check('Exchange', 'i'))
+        {
+            return view('admins.permissions.no');
+        }
         return view('admins.exchanges.create');
     }
     public function edit($id)
     {
+        if(!Right::check('Exchange', 'u'))
+        {
+            return view('admins.permissions.no');
+        }
         $data['exc'] = DB::table('exchanges')->where('id', $id)->first();
         return view('admins.exchanges.edit', $data);
     }
     public function save(Request $r)
     {
+        if(!Right::check('Exchange', 'i'))
+        {
+            return view('admins.permissions.no');
+        }
         $data = array(
             'currency' => $r->currency,
             'rate' => $r->rate
@@ -44,12 +60,20 @@ class ExchangeController extends Controller
     }
     public function delete(Request $r)
     {
+        if(!Right::check('Exchange', 'd'))
+        {
+            return view('admins.permissions.no');
+        }
         DB::table('exchanges')->where('id', $r->id)->delete();
         $r->session()->flash('sms', 'An exchange has been removed successfully!');
         return redirect('analee-admin/exchange');
     }
     public function update(Request $r)
     {
+        if(!Right::check('Exchange', 'u'))
+        {
+            return view('admins.permissions.no');
+        }
         $data = array(
             'currency' => $r->currency,
             'rate' => $r->rate
